one to many maping, embeeded entity relationship done

+++ GameApp/src/main/java/com/cmpe275/GameApp/Repository/SponsorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.cmpe275.GameApp.Entity.Sponsor;
 
-public interface SponsorRepository extends JpaRepository<Sponsor, Integer> {
+public interface SponsorRepository extends JpaRepository<Sponsor, Long> {
 
 }
